+++ src/restful.js
@@ -66,7 +66,7 @@ class RestFul extends EventEmitter {
             });
 
         } catch (error) {
-            this.emit('error', `RESTful error: ${error}`)
+            this.emit('error', `RESTful Connect error: ${error}`)
         }
     };
 
@@ -136,7 +136,7 @@ class RestFul extends EventEmitter {
                 this.restFulData.plcLevel = data;
                 break;
             default:
-                this.emit('debug', `RESTFul update unknown path: ${path}, data: ${data}`)
+                this.emit('error', `RESTFul update unknown path: ${path}, data: ${data}`)
                 break;
         };
         const emitDebug = this.restFulDebug ? this.emit('debug', `RESTFul update path: ${path}, data: ${JSON.stringify(data, null, 2)}`) : false;