+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.2.4] - (xx.08.2024)
+
+## Changes
+
+- fix [#150](https://github.com/grzegorz914/homebridge-enphase-envoy/issues/150)
+- fix voltage divide
+- correct some error logs
+- cleanup
+
 ## [9.2.3] - (18.08.2024)
 
 ## Changes